Сделаны эндпоинты для Genres и Mpa Исправлена инициализация БД, теперь ее не нужно удалять при каждом запуске приложения Созданы объекты Dto

+++ src/main/java/ru/yandex/practicum/filmorate/dao/GenresStorage.java
@@ -8,5 +8,5 @@ import java.util.Optional;
 public interface GenresStorage {
     List<Genre> findAll();
 
-    Optional<Genre> findById(long id);
+    Optional<Genre> findById(int id);
 }
